HTBHF-2705 Removed the v2 claims endpoint and accompanying code.
