+++ api/src/main/java/uk/gov/dhsc/htbhf/claimant/controller/v3/ClaimControllerV3.java
@@ -1,82 +1,99 @@
 package uk.gov.dhsc.htbhf.claimant.controller.v3;
 
 import io.swagger.annotations.*;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import uk.gov.dhsc.htbhf.claimant.controller.v2.ClaimController;
 import uk.gov.dhsc.htbhf.claimant.converter.ClaimantDTOToClaimantConverter;
 import uk.gov.dhsc.htbhf.claimant.converter.VoucherEntitlementToDTOConverter;
+import uk.gov.dhsc.htbhf.claimant.entity.Claimant;
 import uk.gov.dhsc.htbhf.claimant.model.ClaimResultDTO;
-import uk.gov.dhsc.htbhf.claimant.model.v2.AddressDTO;
-import uk.gov.dhsc.htbhf.claimant.model.v2.ClaimDTO;
-import uk.gov.dhsc.htbhf.claimant.model.v2.ClaimantDTO;
-import uk.gov.dhsc.htbhf.claimant.model.v3.AddressDTOV3;
+import uk.gov.dhsc.htbhf.claimant.model.ClaimStatus;
+import uk.gov.dhsc.htbhf.claimant.model.VoucherEntitlementDTO;
 import uk.gov.dhsc.htbhf.claimant.model.v3.ClaimDTOV3;
-import uk.gov.dhsc.htbhf.claimant.model.v3.ClaimantDTOV3;
+import uk.gov.dhsc.htbhf.claimant.service.ClaimRequest;
+import uk.gov.dhsc.htbhf.claimant.service.ClaimResult;
 import uk.gov.dhsc.htbhf.claimant.service.claim.ClaimService;
 import uk.gov.dhsc.htbhf.errorhandler.ErrorResponse;
 
+import java.util.Map;
 import javax.validation.Valid;
 
+import static org.apache.commons.lang3.BooleanUtils.isTrue;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 
 @RestController
 @RequestMapping(value = "/v3/claims", consumes = APPLICATION_JSON_VALUE, produces = APPLICATION_JSON_VALUE)
+@RequiredArgsConstructor
 @Slf4j
 @Api(description = "Endpoints for dealing with claims, e.g. persisting a claim.")
 public class ClaimControllerV3 {
 
-    private final ClaimController claimController;
+    private final ClaimService claimService;
+    private final ClaimantDTOToClaimantConverter claimantConverter;
+    private final VoucherEntitlementToDTOConverter voucherConverter;
 
-    public ClaimControllerV3(ClaimService v3ClaimService,
-                             ClaimantDTOToClaimantConverter claimantConverter,
-                             VoucherEntitlementToDTOConverter voucherConverter) {
-        //TODO MRS 28/11/2019: This is a temporary (ugly) work around to make sure we reuse the existing code without duplication, it will be replaced
-        claimController = new ClaimController(v3ClaimService, claimantConverter, voucherConverter);
-    }
+    private final Map<ClaimStatus, HttpStatus> statusMap = Map.of(
+            ClaimStatus.NEW, HttpStatus.CREATED,
+            ClaimStatus.PENDING, HttpStatus.OK,
+            ClaimStatus.ACTIVE, HttpStatus.OK,
+            ClaimStatus.PENDING_EXPIRY, HttpStatus.OK,
+            ClaimStatus.EXPIRED, HttpStatus.OK,
+            ClaimStatus.REJECTED, HttpStatus.OK,
+            ClaimStatus.ERROR, HttpStatus.INTERNAL_SERVER_ERROR
+    );
 
     @PostMapping
-    @ApiOperation("Create or update a V3 claim.")
+    @ApiOperation("Create a claim.")
     @ApiResponses({@ApiResponse(code = 400, message = "Bad request", response = ErrorResponse.class)})
-    public ResponseEntity<ClaimResultDTO> createOrUpdateClaimV3(@RequestBody @Valid @ApiParam("The claim to persist") ClaimDTOV3 claimDTOV3) {
-        log.debug("Received V3 claim");
-        return claimController.createOrUpdateClaim(buildClaimDTO(claimDTOV3));
+    public ResponseEntity<ClaimResultDTO> createOrUpdateClaimV3(@RequestBody @Valid @ApiParam("The claim to persist") ClaimDTOV3 claimDTO) {
+        log.debug("Received claim");
+        Claimant claimant = claimantConverter.convert(claimDTO.getClaimant());
+        ClaimRequest claimRequest = ClaimRequest.builder()
+                .claimant(claimant)
+                .deviceFingerprint(claimDTO.getDeviceFingerprint())
+                .webUIVersion(claimDTO.getWebUIVersion())
+                .build();
+        ClaimResult result = claimService.createOrUpdateClaim(claimRequest);
+
+        return createResponse(result);
     }
 
-    private ClaimDTO buildClaimDTO(ClaimDTOV3 v3Claim) {
-        return ClaimDTO.builder()
-                .deviceFingerprint(v3Claim.getDeviceFingerprint())
-                .webUIVersion(v3Claim.getWebUIVersion())
-                .claimant(buildClaimantDTO(v3Claim.getClaimant()))
+    private ResponseEntity<ClaimResultDTO> createResponse(ClaimResult result) {
+        ClaimStatus claimStatus = result.getClaim().getClaimStatus();
+        HttpStatus statusCode = getHttpStatus(result);
+        VoucherEntitlementDTO entitlement = getEntitlement(result);
+        ClaimResultDTO body = ClaimResultDTO.builder()
+                .claimStatus(claimStatus)
+                .eligibilityStatus(result.getClaim().getEligibilityStatus())
+                .voucherEntitlement(entitlement)
+                .claimUpdated(result.getClaimUpdated())
+                .updatedFields(result.getUpdatedFields())
+                .verificationResult(result.getVerificationResult())
                 .build();
+        return new ResponseEntity<>(body, statusCode);
     }
 
-    private ClaimantDTO buildClaimantDTO(ClaimantDTOV3 v3Claimant) {
-        return ClaimantDTO.builder()
-                .firstName(v3Claimant.getFirstName())
-                .lastName(v3Claimant.getLastName())
-                .nino(v3Claimant.getNino())
-                .dateOfBirth(v3Claimant.getDateOfBirth())
-                .emailAddress(v3Claimant.getEmailAddress())
-                .phoneNumber(v3Claimant.getPhoneNumber())
-                .address(buildAddressDTO(v3Claimant.getAddress()))
-                .expectedDeliveryDate(v3Claimant.getExpectedDeliveryDate())
-                .childrenDob(v3Claimant.getChildrenDob())
-                .build();
+    private VoucherEntitlementDTO getEntitlement(ClaimResult result) {
+        return result.getVoucherEntitlement().map(voucherConverter::convert).orElse(null);
     }
 
-    private AddressDTO buildAddressDTO(AddressDTOV3 v3Address) {
-        return AddressDTO.builder()
-                .addressLine1(v3Address.getAddressLine1())
-                .addressLine2(v3Address.getAddressLine2())
-                .townOrCity(v3Address.getTownOrCity())
-                .county(v3Address.getCounty())
-                .postcode(v3Address.getPostcode())
-                .build();
+    private HttpStatus getHttpStatus(ClaimResult claimResult) {
+        if (isTrue(claimResult.getClaimUpdated())) {
+            return HttpStatus.OK;
+        }
+        ClaimStatus claimStatus = claimResult.getClaim().getClaimStatus();
+        HttpStatus statusCode = statusMap.get(claimStatus);
+        if (statusCode == null) {
+            log.warn("claim status without HttpStatus: {}", claimStatus);
+            return HttpStatus.INTERNAL_SERVER_ERROR;
+        }
+        return statusCode;
     }
 
 }
