+++ api/src/main/java/uk/gov/dhsc/htbhf/claimant/converter/ClaimantDTOToClaimantConverter.java
@@ -3,12 +3,12 @@ package uk.gov.dhsc.htbhf.claimant.converter;
 import org.springframework.stereotype.Component;
 import org.springframework.util.Assert;
 import uk.gov.dhsc.htbhf.claimant.entity.Claimant;
-import uk.gov.dhsc.htbhf.claimant.model.v2.ClaimantDTO;
+import uk.gov.dhsc.htbhf.claimant.model.v3.ClaimantDTOV3;
 
 import static org.apache.commons.lang3.StringUtils.upperCase;
 
 /**
- * Converts a {@link ClaimantDTO} into a {@link Claimant}.
+ * Converts a {@link ClaimantDTOV3} into a {@link Claimant}.
  */
 @Component
 public class ClaimantDTOToClaimantConverter {
@@ -19,7 +19,7 @@ public class ClaimantDTOToClaimantConverter {
         this.addressConverter = addressConverter;
     }
 
-    public Claimant convert(ClaimantDTO claimant) {
+    public Claimant convert(ClaimantDTOV3 claimant) {
         Assert.notNull(claimant, "source ClaimantDTO must not be null");
         return Claimant.builder()
                 .firstName(claimant.getFirstName())
