+++ api/src/main/java/uk/gov/dhsc/htbhf/claimant/factory/CardRequestFactory.java
@@ -7,7 +7,7 @@ import uk.gov.dhsc.htbhf.claimant.converter.AddressToAddressDTOConverter;
 import uk.gov.dhsc.htbhf.claimant.entity.Claim;
 import uk.gov.dhsc.htbhf.claimant.entity.Claimant;
 import uk.gov.dhsc.htbhf.claimant.model.card.CardRequest;
-import uk.gov.dhsc.htbhf.claimant.model.v2.AddressDTO;
+import uk.gov.dhsc.htbhf.claimant.model.v3.AddressDTOV3;
 
 /**
  * Creates a {@link CardRequest} instance from a given {@link Claim}.
@@ -21,7 +21,7 @@ public class CardRequestFactory {
     public CardRequest createCardRequest(Claim claim) {
         Assert.notNull(claim, "Claim must not be null");
         Claimant claimant = claim.getClaimant();
-        AddressDTO address = addressConverter.convert(claimant.getAddress());
+        AddressDTOV3 address = addressConverter.convert(claimant.getAddress());
         return CardRequest.builder()
                 .address(address)
                 .firstName(claimant.getFirstName())
