+++ api/src/test/java/uk/gov/dhsc/htbhf/claimant/creator/model/ClaimantInfo.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import uk.gov.dhsc.htbhf.claimant.model.v2.AddressDTO;
+import uk.gov.dhsc.htbhf.claimant.model.v3.AddressDTOV3;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -28,7 +28,7 @@ public class ClaimantInfo {
     @JsonProperty("expectedDeliveryDate")
     LocalDate expectedDeliveryDate;
     @JsonProperty("address")
-    AddressDTO addressDTO;
+    AddressDTOV3 addressDTO;
     @JsonProperty("childrenInfo")
     List<ChildInfo> childrenInfo;
 }
